+++ src/count_min_sketch.h
@@ -10,6 +10,11 @@
 
 #include <inttypes.h>       /* PRIu64 */
 
+/* https://gcc.gnu.org/onlinedocs/gcc/Alternate-Keywords.html#Alternate-Keywords */
+#ifndef __GNUC__
+#define __inline__ inline
+#endif
+
 /* hashing function type */
 typedef uint64_t* (*cms_hash_function) (int num_hashes, char *key);
 
@@ -25,12 +30,12 @@ typedef struct {
 
 
 int cms_init_alt(CountMinSketch *cms, int width, int depth, cms_hash_function hash_function);
-static int __inline__  cms_init(CountMinSketch *cms, int width, int depth) {
+static __inline__ int cms_init(CountMinSketch *cms, int width, int depth) {
     return cms_init_alt(cms, width, depth, NULL);
 }
 
 int cms_init_optimal_alt(CountMinSketch *cms, double error_rate, double confidence, cms_hash_function hash_function);
-static int __inline__  cms_init_optimal(CountMinSketch *cms, float error_rate, float confidence) {
+static __inline__ int cms_init_optimal(CountMinSketch *cms, float error_rate, float confidence) {
     return cms_init_optimal_alt(cms, error_rate, confidence, NULL);
 }
 
@@ -43,7 +48,7 @@ int cms_add(CountMinSketch *cms, char* key);
 int cms_remove(CountMinSketch *cms, char* key);
 
 int cms_check(CountMinSketch *cms, char* key);
-static int __inline__ cms_check_min(CountMinSketch *cms, char* key) {
+static __inline__ int cms_check_min(CountMinSketch *cms, char* key) {
     return cms_check(cms, key);
 }
 
@@ -51,8 +56,10 @@ int cms_check_mean(CountMinSketch *cms, char* key);
 
 
 // TODO: add additional functionality
-// uint64_t* cms_get_hashes(CountMinSketch *cms);
-// uint64_t* cms_get_hashes_alt(CountMinSketch *cms, int num_hashes);
+uint64_t* cms_get_hashes_alt(CountMinSketch *cms, int num_hashes, char* key);
+static __inline__ uint64_t* cms_get_hashes(CountMinSketch *cms, char* key) {
+    return cms_get_hashes_alt(cms, cms->depth, key);
+}
 
 #define CMS_SUCCESS 0